Rename utils package argon2
    utils was not explicit enough when using package, argon2 describes exactly what this package does.
