+++ argon2/crypto.go
@@ -1,4 +1,4 @@
-package utils
+package argon2
 
 import (
 	"crypto/rand"
